+++ packages/dialog/index.js
@@ -1,0 +1,8 @@
+import EsDialog from './src/dialog'
+
+EsDialog.install = Vue => Vue.component(EsDialog.name, EsDialog);
+if (typeof window !== 'undefined' && window.Vue) {
+  window.Vue.use(EsDialog);
+}
+
+export default EsDialog;